app/requirements.txt
requests
beautifulsoup4
pytest
boto3
pydantic
feedparser
python-dotenv